Create mixin classes to reuse functionality
